fix: Update SEO setting on the test preview page
